Add support for one-off functions in quickjs
